close files before copying/removing them
WARNING: in save_project_as() I'm pretty sure the database file is copyied before it is closed, on purpose (to avoit commiting modifications to the older file), but nothing ensures all file cache are flushed. Thus the database may end up corrupted, and it's possible that on windows the file copy will be refused.