interactive GUI for iterations (#186)
allwos to specify which parameters will be iterated, and which ones will
be connected to an input filter and the database.
Such nodes are insterted and connected automatically. Lists in iterated
processes/pipelines which are connected to an input filter are also
added a Reduce node to transform files into lists as they expect.