fixed history graphs for data (#236)
Now graphs seem quite complete, including when temp intermediate values
are used (with a little risk of error, however).
But I had to disable the history cleaning for data and orphan bricks for
now, because the notion of obsolete/orphan has chnaged, and it was
erasing things that are used in the indirect history of data.