refactoring and documentation for data history (#236)
The implementation doesn't use Process instances but a light wrapper,
and a pipeline (with fake processes) is only built at the end, in
data_history_pipeline(). Other functions are Process-agnostic.
Docstrings have been added, and a simpler function,
get_data_history_bricks() whic may be used in the history cleaning
functions (which are currently disabled in pipeline runs).